Add approval functionality for user management and section updates

+++ backend/resources/section_resource.py
@@ -1,8 +1,9 @@
 from flask import request, jsonify
 from flask_restful import Resource, marshal, marshal_with, fields, reqparse
-from services import SectionService as service
+from services import SectionService as service, RequestService
 from .resource_utils import validate_date
 from .marshal_fields import section_fields as marshal_fields
+from flask_security import current_user
 
 parser = reqparse.RequestParser()
 parser.add_argument('name', type=str, required=True, help='Name of the section is required')
@@ -14,7 +15,6 @@ class SectionResource(Resource):
         item = service.get_by_id(id)
         return item
     
-    @marshal_with(marshal_fields)
     def put(self, id):
         item = service.get_by_id(id)
         if not item:
@@ -22,11 +22,23 @@ class SectionResource(Resource):
         # args = self.parser.parse_args()
         args = parser.parse_args()
         args['id'] = id
+
+        if (current_user.has_role('manager')):
+            RequestService.create({
+                "data": args,
+                "status": "created",
+                "type": "section_update",
+                "user_id": current_user.id
+            })
+            return {"message": "Update request created and sent for approval"}, 202
+
+
+
         item = service.update(args)
-        return item, 200
+        return marshal(item, marshal_fields), 200
         # service.update(request.json)
 
-    @marshal_with(marshal_fields)
+
     def patch(self, id):
         item = service.get_by_id(id)
         if not item:
@@ -42,13 +54,31 @@ class SectionResource(Resource):
             return {"message": str(e)}, 400
 
         data['id'] = id
+        if (current_user.has_role('manager')):
+            RequestService.create({
+                "data": data,
+                "status": "created",
+                "type": "patch_section",
+                "user_id": current_user.id
+            })
+            return {"message": "Update request created and sent for approval"}, 202
         item = service.update(data)
-        return item, 200
+        return marshal(item, marshal_fields), 200
 
     def delete(self, id):
         item = service.get_by_id(id)
         if not item:
             return {"message": f"Product with id {id} not found"}, 404
+        
+        if (current_user.has_role('manager')):
+            RequestService.create({
+                "data": {"id": id},
+                "status": "created",
+                "type": "delete_section",
+                "user_id": current_user.id
+            })
+            return {"message": "Update request created and sent for approval"}, 202
+        
         message = service.delete(id)
         return message, 200
 
@@ -62,5 +92,15 @@ class SectionListResource(Resource):
     @marshal_with(marshal_fields)
     def post(self):
         args = parser.parse_args()
+
+        if (current_user.has_role('manager')):
+            RequestService.create({
+                "data": args,
+                "status": "created",
+                "type": "post_section",
+                "user_id": current_user.id
+            })
+            return {"message": "Update request created and sent for approval"}, 202
+        
         item = service.create(args)
         return item, 201
